refactor: extract quick switch settings component
